requirements.txt
djangorestframework==3.7.7
Django==2.0
